requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
black==23.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
Django==4.2.5
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
gunicorn==21.2.0
h11==0.14.0
homebrew==0.2.5
httpcore==0.18.0
httpx==0.25.0
icalendar==5.0.7
idna==3.4
iniconfig==2.0.0
logconfig==0.4.0
logutils==0.3.5
mypy-extensions==1.0.0
numpy==1.25.2
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyCLI==2.0.3
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pytest==7.4.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.0
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.27.0
subcommands==3.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.12
uvicorn==0.23.2
